chore!: remove the app from the rdk (#230)
