--- app/evm/codec/codec.go
@@ -1,41 +1,0 @@
-//go:build evm
-// +build evm
-
-package codec
-
-import (
-	"github.com/cosmos/cosmos-sdk/codec"
-	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
-	"github.com/cosmos/cosmos-sdk/std"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/types/module"
-
-	cryptocodec "github.com/evmos/evmos/v12/crypto/codec"
-	ethermint "github.com/evmos/evmos/v12/types"
-
-	"github.com/dymensionxyz/dymension-rdk/app/params"
-)
-
-// MakeEncodingConfig creates a new EncodingConfig with all modules registered
-func MakeEncodingConfig(mb module.BasicManager) params.EncodingConfig {
-	encodingConfig := params.MakeEncodingConfig()
-	RegisterLegacyAminoCodec(encodingConfig.Amino)
-	RegisterInterfaces(encodingConfig.InterfaceRegistry)
-	mb.RegisterLegacyAminoCodec(encodingConfig.Amino)
-	mb.RegisterInterfaces(encodingConfig.InterfaceRegistry)
-	return encodingConfig
-}
-
-// RegisterLegacyAminoCodec registers Interfaces from types, crypto, and SDK std.
-func RegisterLegacyAminoCodec(cdc *codec.LegacyAmino) {
-	sdk.RegisterLegacyAminoCodec(cdc)
-	cryptocodec.RegisterCrypto(cdc)
-	codec.RegisterEvidences(cdc)
-}
-
-// RegisterInterfaces registers Interfaces from types, crypto, and SDK std.
-func RegisterInterfaces(interfaceRegistry codectypes.InterfaceRegistry) {
-	std.RegisterInterfaces(interfaceRegistry)
-	cryptocodec.RegisterInterfaces(interfaceRegistry)
-	ethermint.RegisterInterfaces(interfaceRegistry)
-}
