--- app/evm/flags/flags.go
@@ -1,33 +1,0 @@
-//go:build evm
-// +build evm
-
-package flags
-
-// JSON-RPC flags
-const (
-	JSONRPCEnable              = "json-rpc.enable"
-	JSONRPCAPI                 = "json-rpc.api"
-	JSONRPCAddress             = "json-rpc.address"
-	JSONWsAddress              = "json-rpc.ws-address"
-	JSONRPCGasCap              = "json-rpc.gas-cap"
-	JSONRPCEVMTimeout          = "json-rpc.evm-timeout"
-	JSONRPCTxFeeCap            = "json-rpc.txfee-cap"
-	JSONRPCFilterCap           = "json-rpc.filter-cap"
-	JSONRPCLogsCap             = "json-rpc.logs-cap"
-	JSONRPCBlockRangeCap       = "json-rpc.block-range-cap"
-	JSONRPCHTTPTimeout         = "json-rpc.http-timeout"
-	JSONRPCHTTPIdleTimeout     = "json-rpc.http-idle-timeout"
-	JSONRPCAllowUnprotectedTxs = "json-rpc.allow-unprotected-txs"
-	JSONRPCMaxOpenConnections  = "json-rpc.max-open-connections"
-	JSONRPCEnableIndexer       = "json-rpc.enable-indexer"
-	// JSONRPCEnableMetrics enables EVM RPC metrics server.
-	// Set to `metrics` which is hardcoded flag from go-ethereum.
-	// https://github.com/ethereum/go-ethereum/blob/master/metrics/metrics.go#L35-L55
-	JSONRPCEnableMetrics = "metrics"
-)
-
-// EVM flags
-const (
-	EVMTracer         = "evm.tracer"
-	EVMMaxTxGasWanted = "evm.max-tx-gas-wanted"
-)
