--- app/upgrades.go
@@ -1,40 +1,0 @@
-package app
-
-import (
-	storetypes "github.com/cosmos/cosmos-sdk/store/types"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/types/module"
-	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
-	epochtypes "github.com/dymensionxyz/dymension-rdk/x/epochs/types"
-	claimstypes "github.com/evmos/evmos/v12/x/claims/types"
-)
-
-// UpgradeName defines the on-chain upgrade name for the sample simap upgrade from v045 to v046.
-//
-// NOTE: This upgrade defines a reference implementation of what an upgrade could look like
-// when an application is migrating from Cosmos SDK version v0.45.x to v0.46.x.
-const UpgradeName = "v045-to-v046"
-
-func (app App) RegisterUpgradeHandlers() {
-	app.UpgradeKeeper.SetUpgradeHandler(UpgradeName,
-		func(ctx sdk.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
-			return app.mm.RunMigrations(ctx, app.configurator, fromVM)
-		})
-
-	upgradeInfo, err := app.UpgradeKeeper.ReadUpgradeInfoFromDisk()
-	if err != nil {
-		panic(err)
-	}
-
-	if upgradeInfo.Name == UpgradeName && !app.UpgradeKeeper.IsSkipHeight(upgradeInfo.Height) {
-		storeUpgrades := storetypes.StoreUpgrades{
-			Added: []string{
-				claimstypes.StoreKey,
-				epochtypes.StoreKey,
-			},
-		}
-
-		// configure store loader that checks if version == upgradeHeight and applies store upgrades
-		app.SetStoreLoader(upgradetypes.UpgradeStoreLoader(upgradeInfo.Height, &storeUpgrades))
-	}
-}
