--- scripts/ibc/hub.json
@@ -1,17 +1,0 @@
-{
-  "type": "cosmos",
-  "value": {
-    "key": "relayer-hub-key",
-    "chain-id": "dymension_100-1",
-    "rpc-addr": "tcp://127.0.0.1:36657",
-    "account-prefix": "dym",
-    "keyring-backend": "test",
-    "gas-adjustment": 1.2,
-    "gas-prices": "0.25udym",
-    "debug": true,
-    "timeout": "10s",
-    "output-format": "json",
-    "sign-mode": "direct",
-    "client-type": "07-tendermint"
-  }
-}
