--- scripts/ibc/rollapp.json
@@ -1,17 +1,0 @@
-{
-  "type": "cosmos",
-  "value": {
-    "key": "relayer-rollapp-key",
-    "chain-id": "rollappx",
-    "rpc-addr": "http://localhost:26657",
-    "account-prefix": "rol",
-    "keyring-backend": "test",
-    "gas-adjustment": 1.2,
-    "gas-prices": "0.0urax",
-    "debug": true,
-    "timeout": "10s",
-    "output-format": "json",
-    "sign-mode": "direct",
-    "client-type": "01-dymint"
-  }
-}
