--- scripts/ibc/run_relayer.sh
@@ -1,5 +1,0 @@
-BASEDIR=$(dirname "$0")
-. "$BASEDIR"/../shared.sh
-
-echo '# ------------------------------ run the ibc relayer ----------------------------- #'
-rly start "$RELAYER_PATH" --debug-addr ""
