--- scripts/run_rollapp.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-BASEDIR=$(dirname "$0")
-. "$BASEDIR"/shared.sh
-
-$EXECUTABLE start
-  
-
