--- scripts/settlement/register_sequencer_to_hub.sh
@@ -1,15 +1,0 @@
-#!/bin/bash
-BASEDIR=$(dirname "$0")
-. "$BASEDIR"/../shared.sh
-
-#Register Sequencer
-DESCRIPTION="{\"Moniker\":\"$MONIKER\",\"Identity\":\"\",\"Website\":\"\",\"SecurityContact\":\"\",\"Details\":\"\"}";
-SEQ_PUB_KEY="$($EXECUTABLE dymint show-sequencer --home $ROLLAPP_CHAIN_DIR)"
-
-$SETTLEMENT_EXECUTABLE tx sequencer create-sequencer "$SEQ_PUB_KEY" "$ROLLAPP_CHAIN_ID" "$DESCRIPTION" \
-  --from "$KEY_NAME_DYM" \
-  --chain-id "$SETTLEMENT_CHAIN_ID" \
-  --keyring-dir "$KEYRING_PATH" \
-  --keyring-backend test \
-  --broadcast-mode block \
-  --node "$SETTLEMENT_RPC"
